requirements.txt
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
wandb~=0.19.6

numpy~=1.19.2
pillow~=11.0.0
requests~=2.32.3
beautifulsoup4~=4.12.3
scipy~=1.13.1
tifffile~=2024.8.30
rasterio~=1.4.3